Add http client to hub to handle the task system
Certain HTTP actions from the Pulp API returns a 202, requiring us to parse the
task system to verify the task status.

Abstract the HTTP client to handle tasks as part of its functionality. This
removes the need to check if a specific endpoint will return a 202 or not.

The `requestCommon` function from the framework doesn't provide the status code
in its response.

Unlike modifying the existing `requestCommon` from the
framework, which would necessitate changes in multiple files, we've introduced
a dedicated HTTP client to address HUB-specific requirements.

Also, move nameKeyFn and idKeyFn to a common location. Since it has been use in
several projects.

See: https://pulp.plan.io/issues/3166